refactor: add service layer (#736)
